requirements.txt
streamlit>=1.34.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
chromadb>=0.4.22
exa-py>=1.0.0
pypdf>=3.17.0
python-docx>=1.1.0
openpyxl>=3.1.2
tiktoken>=0.5.2
python-dotenv>=1.0.0

